get-response/requirements.txt
anyio==3.7.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
fastapi==0.99.1
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
googleapis-common-protos==1.59.1
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.10
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
unicorn==2.0.1.post1
urllib3==1.26.16
